Move this repository from Github to Clyde
We decided to make the repository private for increased security. The cheapest option seemed to be to just self-host the repo on this server, rather than paying $$$ to Github, Inc.

This commit prepares the server environment with a shared git user. The repository itself lives somewhere in /srv/git and should be included in the clydesrv.tar.gz backup.

Note that the amended README instructions haven't been tested.