Update Let's Encrypt: wildcard and backups
The backup code jumps through hoops in order to minimise the number of files included in the backup. Whether this in fact makes sense is yet to be determined. Just dumping the entire /etc/letsencrypt dir into the backup would not be a big deal either, as that's only a couple of dozen kilobytes large anyway.