Streamline and accelerate the installation
The number of user interactions required is cut to the absolute minimum (the secret key passphrase). This saves time and nerves.

Additionally, the Perl modules are now installed without testing (seeing as they're by and large already very well tested anyway). Together with perlbrew's parallel building which was already in place, this cuts the installation time almost by half.