More stringent limit for attachment size
After a recent case of severe abuse, it is clear that the limit was far too generous before.

In fact, with use of @skgb.de addresses apparently becoming very rare in future, about a 1 MiB limit may be quite enough. However, should @skgb.de addresses remain the primary mode of communication with and within the board, then the limit should be in the range 2–4 in order to accommodate special circumstances, where unusually large attachments may be expected and necessary.